# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertDraftView": {
        "properties": {
          "alertId": {
            "format": "uuid",
            "type": "string"
          },
          "approvedMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "criticalFields": {
            "items": {
              "$ref": "#/components/schemas/AlertFieldConfirmationView"
            },
            "type": "array"
          },
          "draftVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "location": {
            "type": "string"
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/AlertRecipientSelectionView"
            },
            "type": "array"
          },
          "sbar": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AlertSbarDraft"
              },
              {
                "type": "null"
              }
            ]
          },
          "simulationPatientReference": {
            "type": "string"
          },
          "sourceText": {
            "type": [
              "null",
              "string"
            ]
          },
          "sourceType": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "urgencyLabel": {
            "type": "string"
          }
        },
        "required": [
          "alertId",
          "approvedMessage",
          "criticalFields",
          "draftVersion",
          "location",
          "recipients",
          "sbar",
          "simulationPatientReference",
          "sourceText",
          "sourceType",
          "state",
          "urgencyLabel"
        ],
        "type": "object"
      },
      "AlertFieldConfirmationView": {
        "properties": {
          "alertVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "confirmedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "confirmedByUserId": {
            "format": "uuid",
            "type": "string"
          },
          "fieldId": {
            "type": "string"
          },
          "normalizedValue": {
            "type": "string"
          },
          "originalValue": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "unit": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "alertVersion",
          "confirmedAtUtc",
          "confirmedByUserId",
          "fieldId",
          "normalizedValue",
          "originalValue",
          "status",
          "unit"
        ],
        "type": "object"
      },
...
      "AlertRecipientInput": {
        "properties": {
          "channel": {
            "type": [
              "null",
              "string"
            ]
          },
          "directoryRevision": {
            "type": [
              "null",
              "string"
            ]
          },
          "practitionerId": {
            "format": "uuid",
            "type": "string"
          },
          "practitionerRoleId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "channel",
          "directoryRevision",
          "practitionerId",
          "practitionerRoleId"
        ],
        "type": "object"
      },
      "AlertRecipientSelectionView": {
        "properties": {
          "channel": {
            "type": "string"
          },
          "directoryRevision": {
            "type": "string"
          },
          "directorySourceUpdatedAtUtc": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "onCallSnapshot": {
            "type": [
              "null",
              "string"
            ]
          },
          "practitionerId": {
            "format": "uuid",
            "type": "string"
          },
          "practitionerRoleId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "selectedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "selectionSource": {
            "type": "string"
          }
        },
        "required": [
          "channel",
          "directoryRevision",
          "directorySourceUpdatedAtUtc",
          "onCallSnapshot",
          "practitionerId",
          "practitionerRoleId",
          "selectedAtUtc",
          "selectionSource"
        ],
        "type": "object"
      },
...
      "AlertSbarDraft": {
        "properties": {
          "assessment": {
            "type": [
              "null",
              "string"
            ]
          },
          "background": {
            "type": [
              "null",
              "string"
            ]
          },
          "recommendation": {
            "type": [
              "null",
              "string"
            ]
          },
          "situation": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "assessment",
          "background",
          "recommendation",
          "situation"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "RecipientResponseResult": {
        "properties": {
          "acknowledgedAtUtc": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "alertId": {
            "format": "uuid",
            "type": "string"
          },
          "callUnitRequestedAtUtc": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "confirmedVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "reasonCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "replayed": {
            "type": "boolean"
          },
          "responseType": {
            "type": "string"
          },
          "responsibilityAcceptedAtUtc": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "terminalDisposition": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "acknowledgedAtUtc",
          "alertId",
          "callUnitRequestedAtUtc",
          "confirmedVersion",
          "reasonCode",
          "replayed",
          "responseType",
          "responsibilityAcceptedAtUtc",
          "terminalDisposition"
        ],
        "type": "object"
      },
      "RecordRecipientResponseRequest": {
        "properties": {
          "expectedVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "reasonCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "responseType": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "expectedVersion",
          "responseType"
        ],
        "type": "object"
      },
      "ReplaceAlertRecipientsRequest": {
        "properties": {
          "expectedVersion": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/AlertRecipientInput"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "expectedVersion",
          "recipients"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Simulation-only contract. It is not a hospital or clinical API.",
    "title": "Critical Alerts Simulation API",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/alerts/{alertId}/recipients": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "alertId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ReplaceAlertRecipientsRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertDraftView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "AlertDraftEndpoints"
        ]
      }
    },
...
    "/api/v1/my-alerts/{alertId}/responses": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "alertId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RecordRecipientResponseRequest"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipientResponseResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "tags": [
          "RecipientResponseEndpoints"
        ]
      }
    }
  },
...
}